requirements.txt
pymongo
requests
xmltodict
apache-airflow-providers-mongo